patch from div0 that enables printf format warnings in gcc patch from div0 that fixes a lot of printf format warnings (some quite bad)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6694 d7cf8633-e32d-0410-b094-e92efae38249

+++ r_sky.c
@@ -138,7 +138,7 @@ int R_SetSkyBox(const char *sky)
 
 	if (strlen(sky) > 1000)
 	{
-		Con_Printf("sky name too long (%i, max is 1000)\n", strlen(sky));
+		Con_Printf("sky name too long (%i, max is 1000)\n", (int)strlen(sky));
 		return false;
 	}
